+++ metplus/wrappers/wavelet_stat_wrapper.py
@@ -176,5 +176,6 @@ class WaveletStatWrapper(CompareGriddedWrapper):
             })
 
         self.add_met_config(name='output_prefix', data_type='string')
-
+        # skip RuntimeFreq input file logic - remove once integrated
+        c_dict['FIND_FILES'] = False
         return c_dict